refactor, step 4
