+++ inkstitch/elements/__init__.py
@@ -1,0 +1,5 @@
+from auto_fill import AutoFill
+from fill import Fill
+from stroke import Stroke
+from satin_column import SatinColumn
+from element import EmbroideryElement
