+++ inkstitch/utils/__init__.py
@@ -1,1 +1,2 @@
 from geometry import *
+from cache import cache
