+++ messages.po
@@ -8,7 +8,7 @@ msgid ""
 msgstr ""
 "Project-Id-Version: PACKAGE VERSION\n"
 "Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2018-03-10 22:11-0500\n"
+"POT-Creation-Date: 2018-03-11 21:59-0400\n"
 "PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\n"
 "Last-Translator: FULL NAME <EMAIL@ADDRESS>\n"
 "Language-Team: LANGUAGE <LL@li.org>\n"
@@ -17,144 +17,10 @@ msgstr ""
 "Content-Type: text/plain; charset=CHARSET\n"
 "Content-Transfer-Encoding: 8bit\n"
 
-msgid "Fill"
-msgstr ""
-
-msgid "Manually routed fill stitching"
-msgstr ""
-
-msgid "Angle of lines of stitches"
-msgstr ""
-
-msgid "Flip fill (start right-to-left)"
-msgstr ""
-
-msgid "Spacing between rows"
-msgstr ""
-
-msgid "Maximum fill stitch length"
-msgstr ""
-
-msgid "Stagger rows this many times before repeating"
-msgstr ""
-
-msgid "Auto-Fill"
-msgstr ""
-
-msgid "Automatically routed fill stitching"
-msgstr ""
-
-msgid "Running stitch length (traversal between sections)"
-msgstr ""
-
-msgid "Underlay"
-msgstr ""
-
-msgid "AutoFill Underlay"
-msgstr ""
-
-msgid "Fill angle (default: fill angle + 90 deg)"
-msgstr ""
-
-msgid "Row spacing (default: 3x fill row spacing)"
-msgstr ""
-
-msgid "Max stitch length"
-msgstr ""
-
-msgid "Inset"
-msgstr ""
-
-msgid "Satin stitch along paths"
-msgstr ""
-
-msgid "Running stitch length"
-msgstr ""
-
-msgid "Zig-zag spacing (peak-to-peak)"
-msgstr ""
-
-msgid "Repeats"
-msgstr ""
-
-msgid "Satin Column"
-msgstr ""
-
-msgid "Custom satin column"
-msgstr ""
-
-msgid "Pull compensation"
-msgstr ""
-
-msgid "Contour underlay"
-msgstr ""
-
-msgid "Contour Underlay"
-msgstr ""
-
-msgid "Stitch length"
-msgstr ""
-
-msgid "Contour underlay inset amount"
-msgstr ""
-
-msgid "Center-walk underlay"
-msgstr ""
-
-msgid "Center-Walk Underlay"
-msgstr ""
-
-msgid "Zig-zag underlay"
-msgstr ""
-
-msgid "Zig-zag Underlay"
-msgstr ""
-
-msgid "Zig-Zag spacing (peak-to-peak)"
-msgstr ""
-
-msgid "Inset amount (default: half of contour underlay inset)"
-msgstr ""
-
-msgid ""
-"One or more rails crosses itself, and this is not allowed.  Please split "
-"into multiple satin columns."
-msgstr ""
-
-msgid "satin column: One or more of the rungs doesn't intersect both rails."
-msgstr ""
-
-msgid "Each rail should intersect both rungs once."
-msgstr ""
-
-msgid ""
-"satin column: One or more of the rungs intersects the rails more than once."
-msgstr ""
-
-#, python-format
-msgid "satin column: object %s has a fill (but should not)"
-msgstr ""
-
-#, python-format
-msgid ""
-"satin column: object %(id)s has two paths with an unequal number of points "
-"(%(length1)d and %(length2)d)"
-msgstr ""
-
 msgid ""
 "\n"
 "\n"
 "Seeing a 'no such option' message?  Please restart Inkscape to fix."
-msgstr ""
-
-msgid "No embroiderable paths selected."
-msgstr ""
-
-msgid "No embroiderable paths found in document."
-msgstr ""
-
-msgid ""
-"Tip: use Path -> Object to Path to convert non-paths before embroidering."
 msgstr ""
 
 msgid "These settings will be applied to 1 object."
@@ -240,18 +106,4 @@ msgstr ""
 
 #, python-format
 msgid "Unknown unit: %s"
-msgstr ""
-
-msgid "TRIM after"
-msgstr ""
-
-msgid "Trim thread after this object (for supported machines and file formats)"
-msgstr ""
-
-msgid "STOP after"
-msgstr ""
-
-msgid ""
-"Add STOP instruction after this object (for supported machines and file "
-"formats)"
 msgstr ""
